images/solution/requirements.txt
mysqlclient ~= 1.4
mysql-connector-python-rf